read-http-headers: don't collect unknown headers
After c29d92e6077dbee59a18fa7b410176b0f31205d1 READ-HTTP-HEADERS stopped to
intern unknown headers, but it still collected strings instead.

The function READ-HTTP-HEADERS is specified to return pairs (name . value) where
the name is a keyword. Some clients this alist propagated to HUNCHENTOOT to make
a list of initargs leading to an error that a string is not a valid initarg at
runtime.

This commit changes READ-HTTP-HEADERS to not collect unknown headers at all and
to signal a warning when the first header like this is encountered -- the latter
is to notify the programmer that at least some headers, that they may be
potentially interested in, are ignored.

Fixes #14.